extract default details content
